refactor: add strum derives to error enums and move errors to dedicated modules
Move error types to dedicated error.rs modules for peer, identify,
dialer, and peer-store crates. Add strum::IntoStaticStr derives to all
error enums for metrics integration. Replace manual Display/Error impls
with thiserror derives on NodeBuildError and AddError. Replace manual
LabelValue impl on GossipCheckError with impl_record_error! macro.
Remove redundant #[strum(serialize)] overrides where serialize_all
already produces the correct value.
