+++ crates/net/peer/store/Cargo.toml
@@ -18,6 +18,7 @@ test-utils = ["dep:parking_lot"]
 [dependencies]
 parking_lot = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
+strum = { workspace = true }
 thiserror = { workspace = true }
 auto_impl = { workspace = true }
