+++ crates/node/builder/src/error.rs
@@ -3,7 +3,8 @@
 /// Error during node launch.
 ///
 /// Wraps both protocol-specific errors and infrastructure errors.
-#[derive(Debug, thiserror::Error)]
+#[derive(Debug, thiserror::Error, strum::IntoStaticStr)]
+#[strum(serialize_all = "snake_case")]
 pub enum LaunchError<E: std::error::Error + 'static> {
     /// Protocol-specific error during component build or service spawn.
     #[error("protocol error: {0}")]
@@ -14,7 +15,8 @@ pub enum LaunchError<E: std::error::Error + 'static> {
 }
 
 /// Infrastructure-level errors during node launch.
-#[derive(Debug, thiserror::Error)]
+#[derive(Debug, thiserror::Error, strum::IntoStaticStr)]
+#[strum(serialize_all = "snake_case")]
 pub enum InfrastructureError {
     /// Failed to build gRPC reflection service.
     #[error("gRPC reflection service: {0}")]
