+++ crates/swarm/builder/src/error.rs
@@ -3,7 +3,8 @@
 use thiserror::Error;
 
 /// Error type for Swarm node building and launching.
-#[derive(Debug, Error)]
+#[derive(Debug, Error, strum::IntoStaticStr)]
+#[strum(serialize_all = "snake_case")]
 pub enum SwarmNodeError {
     /// Build error from protocol.
     #[error("build error: {0}")]
