+++ crates/swarm/net/identify/src/lib.rs
@@ -5,6 +5,7 @@
 
 mod behaviour;
 mod config;
+pub mod error;
 mod generated;
 mod handler;
 pub mod metrics;
@@ -12,7 +13,7 @@ mod protocol;
 
 pub use behaviour::{AgentVersions, Behaviour, Event, new_agent_versions};
 pub use config::Config;
-pub use protocol::{Info, PushInfo, UpgradeError};
+pub use protocol::{Info, PushInfo};
 
 use libp2p::swarm::StreamProtocol;
