+++ crates/swarm/net/pseudosettle/src/error.rs
@@ -5,7 +5,6 @@ vertex_net_codec::protocol_error! {
     pub enum PseudosettleError {
         /// Invalid timestamp in acknowledgment.
         #[error("invalid timestamp: {0}")]
-        #[strum(serialize = "invalid_timestamp")]
         InvalidTimestamp(String),
     }
 }
