+++ crates/swarm/net/retrieval/src/error.rs
@@ -5,12 +5,10 @@ vertex_net_codec::protocol_error! {
     pub enum RetrievalError {
         /// Invalid chunk address length.
         #[error("invalid chunk address length: expected 32, got {0}")]
-        #[strum(serialize = "invalid_address_length")]
         InvalidAddressLength(usize),
 
         /// Invalid chunk address encoding.
         #[error("invalid chunk address: {0}")]
-        #[strum(serialize = "invalid_address")]
         InvalidAddress(String),
     }
 }
