+++ crates/swarm/node/Cargo.toml
@@ -66,13 +66,13 @@ vertex-net-peer-store.workspace = true
 bytes.workspace = true
 eyre.workspace = true
 parking_lot.workspace = true
+strum.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 
 ## cli (optional)
 clap = { workspace = true, features = ["derive", "env"], optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
-strum = { workspace = true, optional = true }
 vertex-node-api = { workspace = true, optional = true }
 vertex-swarm-bandwidth-pricing = { workspace = true, optional = true, features = ["cli"] }
 vertex-swarm-localstore = { workspace = true, optional = true }
@@ -88,7 +88,6 @@ std = []
 cli = [
     "dep:clap",
     "dep:serde",
-    "dep:strum",
     "dep:vertex-node-api",
     "dep:vertex-swarm-bandwidth-pricing",
     "dep:vertex-swarm-localstore",
