+++ crates/swarm/peers/peer-manager/Cargo.toml
@@ -36,6 +36,7 @@ libp2p = { workspace = true }
 parking_lot = { workspace = true }
 tracing = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
+strum = { workspace = true }
 thiserror = { workspace = true }
 
 # Concurrency
