+++ crates/swarm/peers/peer/Cargo.toml
@@ -25,6 +25,7 @@ libp2p.workspace = true
 vertex-net-local.workspace = true
 arbitrary = { workspace = true, features = ["derive"] }
 proptest.workspace = true
+strum.workspace = true
 thiserror.workspace = true
 auto_impl = "1.1"
