+++ crates/swarm/topology/src/gossip/error.rs
@@ -1,10 +1,10 @@
 //! Gossip check errors and verification failure types.
 
 use vertex_net_dialer::EnqueueError;
-use vertex_observability::LabelValue;
 
 /// Why a gossiped peer was rejected before verification.
-#[derive(Debug, thiserror::Error)]
+#[derive(Debug, thiserror::Error, strum::IntoStaticStr)]
+#[strum(serialize_all = "snake_case")]
 pub(super) enum GossipCheckError {
     #[error("no multiaddrs")]
     NoMultiaddrs,
@@ -16,15 +16,8 @@ pub(super) enum GossipCheckError {
     Enqueue(#[from] EnqueueError),
 }
 
-impl LabelValue for GossipCheckError {
-    fn label_value(&self) -> &'static str {
-        match self {
-            Self::NoMultiaddrs => "no_multiaddrs",
-            Self::NoPeerId => "no_peer_id",
-            Self::GossiperRateLimited => "gossiper_rate_limited",
-            Self::Enqueue(e) => e.label_value(),
-        }
-    }
+impl GossipCheckError {
+    vertex_metrics::impl_record_error!("topology_gossip_rejected_total");
 }
 
 /// Verification failure after handshake.
@@ -33,4 +26,8 @@ impl LabelValue for GossipCheckError {
 pub(super) enum VerificationFailure {
     /// The gossiped multiaddr is not in the verified peer's address list.
     MultiAddrNotInPeer,
+}
+
+impl VerificationFailure {
+    vertex_metrics::impl_record_error!("topology_gossip_verification_failed_total");
 }
